llvm: Fix source build with modern GCC on Linux.
Modern GCCs on Linux issue a warning that breaks the build that is
disabled with a flag next to `-fpermissive`. This warning doesn't
trigger upstream, but there isn't a point release with this fix.

Also, modern GCCs / linkers will build parts of Polly without PIC and
then complain about linking it into the DSO. However, PIC has very low
cost on Linux and so a simple workaround is just to fully enable a PIC
build. Given that the build already uses both an LLVM and Clang DSO to
reduce the install size, this has minimal impact and mostly fixes the
linking of Polly. It's unclear why this link ever succeeded when
building a bottle, but may reflect system-wide defaults for using PIC.
At least on a Debian testing Linux machine, this is necessary for me to
get the build from source to succeed.